+++ include/shell/commands.h
@@ -59,4 +59,8 @@ void cmd_kill(const char* args);
 void cmd_peek(const char* args);
 void cmd_tlist(const char* args);
 
+// syshw
+void cmd_shutdown(const char* args);
+void cmd_reboot(const char* args);
+
 #endif